File added for coreData
