+++ TAWKIOSTask/Model/UserListModel.swift
@@ -31,42 +31,3 @@ struct UserProfileModel : Codable {
 }
 
 
-
-// 1
-@objc(Userlist)
-// 2
-final class Userlist: NSManagedObject {
-    // 3
-    @NSManaged var id: String
-    @NSManaged var login, type, node_id: String
-    @NSManaged var avatar_url : Data
-
-}
-
-
-extension Userlist {
-    // 4
-    @nonobjc class func fetchRequest() -> NSFetchRequest<Userlist> {
-        return NSFetchRequest<Userlist>(entityName: "Userlist")
-    }
-}
-
-
-// 1
-@objc(Profile)
-// 2
-final class Profile: NSManagedObject {
-    // 3
-    @NSManaged var id: Int
-    @NSManaged var login, type,company, blog, name, note,followers, following: String
-    @NSManaged var avatar_url : Data
-
-}
-
-
-extension Profile {
-    // 4
-    @nonobjc class func fetchRequest() -> NSFetchRequest<Profile> {
-        return NSFetchRequest<Profile>(entityName: "Profile")
-    }
-}
